Auto_Test_Project/requirements.txt
aiohttp==3.9.0
allure-pytest==2.13.2
beautifulsoup4==4.12.2
dingtalkchatbot==1.5.7
loguru==0.7.2
pymysql==1.1.0
pytest==7.4.3
pytest-assume==2.4.3
pytest-ordering==0.6
pytest-rerunfailures==12.0
pytest-xdist==3.4.0
PyYaml==6.0.1
redis==5.0.1
requests==2.31.0
selenium==4.15.2
sqlalchemy==2.0.25
sshtunnel==0.4.0
urllib3==1.26.15

